--- client/README.md
@@ -1,2 +1,0 @@
-# ATAS - Alert Tracking and Acknowledgement System
-Created with ReactJs using typescript
