--- client/package.json
@@ -1,65 +1,0 @@
-{
-  "name": "client",
-  "version": "0.1.0",
-  "private": false,
-  "license": "MIT",
-  "dependencies": {
-    "@chakra-ui/icons": "^1.0.4",
-    "@chakra-ui/react": "^1.1.6",
-    "@emotion/react": "^11.1.4",
-    "@emotion/styled": "^11.0.0",
-    "@glideapps/glide-data-grid": "^2.0.4",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.53",
-    "@types/react-dom": "^16.9.8",
-    "direction": "^1.0.4",
-    "framer-motion": "^3.2.2-rc.1",
-    "lodash": "^4.17.20",
-    "marked": "^2.0.0",
-    "moment": "^2.29.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-redux": "^7.2.2",
-    "react-responsive-carousel": "^3.2.12",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "react-table": "^7.6.3",
-    "redux": "^4.0.5",
-    "styled-components": "^5.2.1",
-    "typescript": "^4.0.3",
-    "web-vitals": "^0.2.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@types/react-redux": "^7.1.16",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/react-table": "^7.0.28"
-  }
-}
