--- client/src/components/Layout.tsx
@@ -1,12 +1,0 @@
-import { Box } from "@chakra-ui/react";
-import React from "react";
-
-const Layout: React.FC = ({ children }) => {
-  return (
-    <Box className="App" paddingLeft="10%" paddingRight="10%" flexGrow={1}>
-      {children}
-    </Box>
-  );
-};
-
-export default Layout;
