--- client/src/components/pages/Homepage.tsx
@@ -1,15 +1,0 @@
-import React, { Component } from "react";
-import Table from "../Alert/Table";
-
-interface Props {}
-interface State {}
-
-export default class Homepage extends Component<Props, State> {
-  render() {
-    return (
-      <div>
-        <Table />
-      </div>
-    );
-  }
-}
