--- client/src/components/pages/index.ts
@@ -1,3 +1,0 @@
-export { default as Homepage } from "./Homepage";
-export { default as Login } from "./Login";
-export { default as Logout } from "./Logout";
