--- client/src/util/consts.ts
@@ -1,1 +1,0 @@
-export const API_DOMAIN = "http://localhost:5000";
