--- client/src/util/gate.ts
@@ -1,9 +1,0 @@
-export const timePromise = (ms: number) => {
-  const timeout = new Promise<undefined>((resolve, reject) => {
-    let wait = setTimeout(() => {
-      clearTimeout(wait);
-      reject("Unable to contact server.");
-    }, ms);
-  });
-  return timeout;
-};
