+++ next-client/pages/index.tsx
@@ -1,6 +1,9 @@
 import { Box } from "@chakra-ui/layout";
+import { SkeletonText } from "@chakra-ui/skeleton";
 import { useRouter } from "next/router";
 import { FunctionComponent, useEffect } from "react";
+import EntryFilter from "../components/entry/EntryFilter";
+import EntryModal from "../components/entry/EntryModal";
 import EntryTable from "../components/entry/EntryTable";
 import { useUser } from "../util/swrFunctions";
 
@@ -12,11 +15,20 @@ const Home: FunctionComponent = () => {
     if (!data && !loading) router.push("/login");
   }, [data, loading]);
 
-  if (!data) return <></>;
+  if (!data)
+    return (
+      <Box paddingTop="1%" paddingLeft="10%" paddingRight="10%">
+        <SkeletonText spacing={6} noOfLines={10}>
+          Placeholder
+        </SkeletonText>
+      </Box>
+    );
 
   return (
-    <Box paddingTop="1%">
+    <Box padding="1%" paddingLeft="10%" paddingRight="10%">
+      <EntryFilter />
       <EntryTable />
+      <EntryModal />
     </Box>
   );
 };
