+++ next-client/util/swrFunctions.tsx
@@ -1,7 +1,4 @@
-import { useContext } from "react";
-import { default as useSWR, default as useSwr } from "swr";
-import { EntryContext } from "../context/EntryContext";
-import IEntry from "../interface/IEntry";
+import { default as useSwr } from "swr";
 import Iuser from "../interface/IUser";
 import { API_DOMAIN } from "./constants";
 
@@ -19,38 +16,6 @@ export const useUser = () => {
   return {
     data: !error ? data : undefined,
     loading: !data && !error,
-    error,
-    mutate,
-  };
-};
-
-interface IEntryReturn {
-  alerts: IEntry[];
-  count: number;
-}
-
-export const useEntries = () => {
-  const context = useContext(EntryContext);
-
-  const CustomFetcher = (
-    input: RequestInfo,
-    init: RequestInit = {
-      credentials: "include",
-      method: "POST",
-      body: JSON.stringify(context.filter),
-      headers: { "Content-Type": "application/json" },
-    }
-  ) => {
-    return fetch(`${API_DOMAIN}${input}`, init).then((response) =>
-      response.json()
-    );
-  };
-
-  const { data, error, mutate } = useSWR<IEntryReturn>("/alert", CustomFetcher);
-
-  return {
-    data: !error ? data : undefined,
-    loading: !error && !data,
     error,
     mutate,
   };
