+++ package.json
@@ -4,7 +4,8 @@
     "concurrently": "^5.3.0"
   },
   "scripts": {
-    "dev": "concurrently \"yarn --cwd client start\" \"yarn --cwd server dev\"",
+    "dev": "concurrently \"yarn --cwd next-client dev\" \"yarn --cwd server dev\"",
     "setup": "yarn  && yarn --cwd client && yarn --cwd server"
-  }
+  },
+  "dependencies": {}
 }
